requirements.txt
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
certifi==2024.6.2
channels==4.1.0
channels-redis==4.2.0
charset-normalizer==3.3.2
Django==4.2.13
django-braces==1.15.0
django-debug-toolbar==3.6.0
django-embed-video==1.4.4
django-redisboard==8.4.0
djangorestframework==3.15.1
idna==3.7
msgpack==1.0.8
pillow==10.3.0
pymemcache==3.5.2
redis==5.0.6
requests==2.32.3
six==1.16.0
sqlparse==0.5.0
typing_extensions==4.12.2
urllib3==2.2.1
